Gate accessibility opt-out: one shared page table, exemptions written down
The gate kept its own page list, so a new page was outside it until someone
remembered to add it — the standard applied to whoever thought of it. Both specs
now read one table (pages.js), which inverts that: a page added for the smoke
tests is held to WCAG 2.1 AA from the same commit, and escaping takes a named
entry in EXEMPT_PAGES rather than an omission nobody can see.

EXEMPT_PAGES is empty and should stay that way. It is a different claim from the
allowlist — "this page is not held to the standard" rather than "this violation
is known and tracked" — so the two are separate rather than one list doing both
jobs. Per-page allowlist entries for pages with nothing to forgive are gone;
absence already means clean.

The smoke suite gains the 17 api-docs pages the gate had been carrying alone.

Also fixes the two findings on the registered-user pages, both the same muted-
copy case as the rest: the dashboard's cities nudge and the settings field help
put link blue against copy it cannot separate itself from by hue.

Part of #5060
