+++ test/e2e/a11y-allowlist.js
@@ -1,47 +1,28 @@
 /**
- * The accessibility gate's per-page allowlist and the helpers that apply it (issue #5060).
+ * The accessibility gate's exemptions, its per-page allowlist, and the helpers that apply them (issue #5060).
  *
- * The allowlist is a work queue, not an exemption list: `issue` is required and must name a real tracked issue, so a
- * reader can tell a known-and-scheduled failure from one nobody has looked at. The Year 1 target (#5053) is an empty
- * list for the homepage and the `/api` docs pages.
+ * Two different escape hatches, both deliberately awkward:
  *
- * Scope an entry as tightly as the violation allows — a bare `{rule}` silences that rule for the whole page, hiding
- * the next instance too. `selector` is matched as a substring of the node's target, so `.au-body` covers
- * `#main > .au-body`.
+ * - EXEMPT_PAGES drops a whole page out of the gate. Coverage is otherwise opt-out — every page in pages.js is
+ *   gated — so this is the only way a page avoids the standard, and it needs a reason someone can argue with.
+ * - A11Y_ALLOWLIST forgives specific known violations on a page that is otherwise gated. It is a work queue, not an
+ *   exemption list: `issue` is required and must name a real tracked issue, so a reader can tell a
+ *   known-and-scheduled failure from one nobody has looked at.
+ *
+ * Scope an allowlist entry as tightly as the violation allows — a bare `{rule}` silences that rule for the whole
+ * page, hiding the next instance too. `selector` is matched as a substring of the node's target, so `.au-body`
+ * covers `#main > .au-body`.
  */
+
+/**
+ * Pages held out of the gate entirely, keyed by path, with the reason. Empty is the goal: a page here is one we have
+ * decided not to hold to WCAG 2.1 AA yet, which is a bigger claim than forgiving a violation.
+ */
+const EXEMPT_PAGES = {
+};
 
 /** Known, tracked violations keyed by page path, as `{rule, selector?, issue, note}`. */
 const A11Y_ALLOWLIST = {
-  '/': [],
-  '/signIn': [],
-  '/signUp': [],
-  '/about': [],
-  '/leaderboard': [],
-  '/routes': [],
-  '/stories': [],
-  '/labelingGuide': [],
-  '/help': [],
-  '/api': [],
-  '/v3/api-docs': [],
-  '/v3/api-docs/labelTypes': [],
-  '/v3/api-docs/cities': [],
-  '/v3/api-docs/labelTags': [],
-  '/v3/api-docs/rawLabels': [],
-  '/v3/api-docs/labelClusters': [],
-  '/v3/api-docs/streets': [],
-  '/v3/api-docs/streetTypes': [],
-  '/v3/api-docs/regions': [],
-  '/v3/api-docs/accessScoreStreets': [],
-  '/v3/api-docs/accessScoreRegions': [],
-  '/v3/api-docs/validations': [],
-  '/v3/api-docs/validation-result-types': [],
-  '/v3/api-docs/labelEdits': [],
-  '/v3/api-docs/user-stats': [],
-  '/v3/api-docs/overall-stats': [],
-  '/v3/api-docs/overall-stats-by-day': [],
-  '/v3/api-docs/aggregate-stats': [],
-  '/v3/api-docs/aggregate-stats-by-day': [],
-  '/gallery': [],
   // No selector on these two: MapboxSearchBox hashes its class names per mount (.mbx0420900a--Input one load,
   // .mbx00a6ef43--Input the next), so there is nothing stable to scope the entry to.
   '/labelMap': [
@@ -52,8 +33,6 @@ const A11Y_ALLOWLIST = {
     {rule: 'aria-required-attr', issue: '#5087',
       note: "MapboxSearchBox's input is role=combobox with no aria-expanded; vendor DOM"},
   ],
-  '/cities': [],
-  '/mobileLanding': [],
 };
 
 /**
@@ -106,4 +85,4 @@ function formatViolations(violations) {
     `${violation.id} [${violation.impact}] ${node.target.join(' ')} — ${violation.help} (${violation.helpUrl})`));
 }
 
-module.exports = {A11Y_ALLOWLIST, partitionViolations, formatViolations};
+module.exports = {EXEMPT_PAGES, A11Y_ALLOWLIST, partitionViolations, formatViolations};
