+++ test/e2e/a11y.spec.js
@@ -1,17 +1,19 @@
 /**
- * Accessibility gate (issue #5060): axe-core over each audited page, failing on any WCAG 2.1 AA violation that
- * a11y-allowlist.js does not already track.
+ * Accessibility gate (issue #5060): axe-core over every page the browser suite covers, failing on any WCAG 2.1 AA
+ * violation that a11y-allowlist.js does not already track.
  *
  * This is the mechanical half of WCAG only — a green run is not a claim of conformance, just that nothing
  * mechanically detectable regressed. The manual half is the checklist in docs/accessibility.md.
  *
- * The page table is a separate list from pages.spec.js's rather than a shared one: pages join this gate as their
- * violations are fixed, so a page not listed here is a page nobody has audited yet. Loading goes through the smoke
- * suite's `loadAndSettle`, so what axe measures is the same settled DOM the error gate measures.
+ * The page list is pages.js, shared with the smoke suite, so coverage here is **opt-out**: a page added there is
+ * gated by default and stays gated unless someone writes it into EXEMPT_PAGES with a reason. Loading goes through
+ * `loadAndSettle`, so what axe measures is the same settled DOM the error gate measures.
  */
 const {test, expect, loadAndSettle} = require('./fixtures');
 const AxeBuilder = require('@axe-core/playwright').default;
-const {partitionViolations, formatViolations} = require('./a11y-allowlist');
+const {PAGES} = require('./pages');
+const {EXEMPT_PAGES, partitionViolations, formatViolations} = require('./a11y-allowlist');
+
 
 // WCAG 2.1 AA, which subsumes 2.0 A and AA. axe's `best-practice` tag is left off: useful advice, but not part of
 // the standard we hold ourselves to, and a gate that fails outside the commitment gets ignored.
@@ -24,48 +26,9 @@ const WCAG_TAGS = ['wcag2a', 'wcag2aa', 'wcag21aa'];
 // our own <iframe> tag out of scope too — the accessible name that costs us is asserted by its own test below.
 const THIRD_PARTY_FRAMES = 'iframe[src*="youtube.com"]';
 
-// Load flags mirror each page's entry in pages.spec.js.
-const PAGES = [
-  // Landing maps are deferred behind util.onFirstInteractionOrIdle (#4486) — without the wait, axe scans a
-  // half-rendered page.
-  {path: '/', mapbox: true, waitFor: (page) => page.waitForFunction(() => window.choropleth && window.deploymentMap)},
-  {path: '/signIn'},
-  {path: '/signUp'},
-  {path: '/about', makeabilityLab: true, mapbox: true},
-  {path: '/leaderboard'},
-  {path: '/routes'},
-  {path: '/stories'},
-  {path: '/labelingGuide'},
-  {path: '/help'},
-  // Every /v3/api-docs page (#5060's Year 1 target is an empty allowlist across the whole reference).
-  {path: '/api'},
-  {path: '/v3/api-docs'},
-  {path: '/v3/api-docs/labelTypes'},
-  {path: '/v3/api-docs/cities'},
-  {path: '/v3/api-docs/labelTags'},
-  {path: '/v3/api-docs/rawLabels'},
-  {path: '/v3/api-docs/labelClusters'},
-  {path: '/v3/api-docs/streets'},
-  {path: '/v3/api-docs/streetTypes'},
-  {path: '/v3/api-docs/regions'},
-  {path: '/v3/api-docs/accessScoreStreets'},
-  {path: '/v3/api-docs/accessScoreRegions'},
-  {path: '/v3/api-docs/validations'},
-  {path: '/v3/api-docs/validation-result-types'},
-  {path: '/v3/api-docs/labelEdits'},
-  {path: '/v3/api-docs/user-stats'},
-  {path: '/v3/api-docs/overall-stats'},
-  {path: '/v3/api-docs/overall-stats-by-day'},
-  {path: '/v3/api-docs/aggregate-stats'},
-  {path: '/v3/api-docs/aggregate-stats-by-day'},
-  {path: '/gallery', loadingOverlay: true},
-  {path: '/labelMap', mapbox: true, loadingOverlay: true},
-  {path: '/routeBuilder', mapbox: true},
-  {path: '/cities', mapbox: true},
-  {path: '/mobileLanding'},
-];
+const GATED_PAGES = PAGES.filter((p) => !(p.path in EXEMPT_PAGES));
 
-for (const p of PAGES) {
+for (const p of GATED_PAGES) {
   test(`a11y: ${p.path} meets WCAG 2.1 AA`, async ({page, context}) => {
     await loadAndSettle(page, context, p);
     const results = await new AxeBuilder({page})
